Test creation of empty Maps (#1)
* Add comments marking projections

* Correct path to shapefiles

* Change plot method to separate save/show methods

Instead of calling `Map.plot()`, you'd call `Map.plot()` and then `Map.save()`

* Add methods to return supported projections/domains

* Add shapefiles to source package

They're now included in the repo and listed in `MANIFEST.in`

* Add tests for empty Map images

`test_map.py` creates empty Map images for each supported projection, and compares that image to the pre-generated image for that projection.

* Specify Matplotlib backend in .travis.yml

Travis CI is failing to create Maps because the default backend is failing due to the lack of a DISPLAY. Setting the backend to GTKAgg fixes this because it doesn't require a X11 display.

* GTKagg needs PyGTK, trying Agg

* Travis CI DEBUG: list images dir

* Travis CI DEBUG

* Fix up assert statement for comparing images

* Travis CI DEBUG

* Travis debug

* Remove -s option for py.test

* Don't create temp dir, just temp files for testing

For testing empty Map images, just create the temp files, not a temp dir. Travis probably doesn't allow writing to whatever directory the test function was attempting to write to.

* Travis debug

* debug

* debug

* debug

* debug

* Use external image diff tool instead of filecmp

filecmp was reporting one of the images to be different on Travis. Instead I am using a new external tool that I am uploading to PyPI, img-percent-diff, that reports the percent area that two images differ.

Asserting that this doesn't go above 0.001 should suffice in asserting they are "the same".

* Remove print debug lines

* Create new test image without 25N lat label

It turns out that on Linux, basemap is *not* labelling 25 deg N, but on Mac OS X it *is*. This is why Travis CI has been reporting a difference between the test image and the image included in the repo.

I am making the image in the repo *without* the 25N label so it'll match that made on Travis CI.