Make oceans dark and land light
This was done by calling `basemap.drawlsmask()` and settings land color to  `#FFFFFF` and ocean color to `(0.4, 0.4, 0.4, 0.9). Then when data is plotted of the land, it remains bright, but data plotted over the ocean is darkened a bit by the dark grey seeping in from underneath.

Also updated test images so tests will still pass.

Also add a `cpc.geoplot.colors` module for storing common lists of colors for plotting.

Also made a script to remake the test images.